New Updater

+++ wiki-updater.sh
@@ -1,9 +1,16 @@
 #!/bin/bash
 
 function UPDATE_WIKI () {
-    sudo cp ./wiki-updater.sh /wiki-updater.old
-#    curl -o ./wiki-updater.sh https://raw.githubusercontent.com/mosesrenegade/sec588-public/master/wiki-updater.sh
-    echo "You may have to run this application two times to update the shell script"
+    curl -o /tmp/wiki-updater.sh https://raw.githubusercontent.com/mosesrenegade/sec588-public/master/wiki-updater.sh
+    sha256sum /opt/wiki/wiki-updater.sh | awk '{ print $1 }' > /opt/wiki/wiki-updater.sh.sha256
+    sha256sum /tmp/wiki-updater.sh | awk '{ print $1 }' > /tmp/wiki-updater.sh.sha256
+    if ! cmp --silent "/opt/wiki/wiki-updater.sh.sha256" "/tmp/wiki-updater.sh.sha256"
+    then
+      cp /opt/wiki/wiki-updater.sh /opt/wiki/wiki-updater.old
+      mv /tmp/wiki-updater.sh /opt/wiki/wiki-updater.sh
+      echo "We have had an update to the updater, we are exiting and will require you to RERUN the application"
+      exit(1)
+    fi
     cd /opt/wiki/sec588-labs-e01
     git reset --hard
     git pull
